refactor tests folders
